% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{country_names}
\alias{country_names}
\title{Íslenskar þýðingar á enskum nöfnum á löndum}
\usage{
country_names()
}
\value{
A tibble with two columns: country (english name), and land (icelandic name)
}
\description{
Íslenskar þýðingar á enskum nöfnum á löndum
}
