% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stjornmalaflokkar.R
\name{flokkar_palette}
\alias{flokkar_palette}
\title{Icelandic political party colour palette}
\usage{
flokkar_palette()
}
\value{
A tibble containing the name of the party, it's abbreviation and it's colour in hex-code
}
\description{
Icelandic political party colour palette
}
