% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{hlutf}
\alias{hlutf}
\title{Format Numbers as Percentages with Icelandic Standards}
\usage{
hlutf(x, accuracy = 0.1, ...)
}
\arguments{
\item{x}{Numeric vector to be formatted as percentage}

\item{accuracy}{A number to round to. E.g. 0.01 shows 2 decimal places.}

\item{...}{Other inputs passed on to scales::percent()}
}
\value{
A character vector formatted as percentage with Icelandic
localisation
}
\description{
Format Numbers as Percentages with Icelandic Standards
}
