% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{label_hlutf}
\alias{label_hlutf}
\title{Create a Labeller to Format Percentages with Icelandic Standards}
\usage{
label_hlutf(accuracy = 0.1, ...)
}
\arguments{
\item{accuracy}{A number to round to. E.g. 0.01 shows 2 decimal places.}

\item{...}{Other inputs passed on to scales::label_percent()}
}
\value{
A labelling function for use with ggplot2 scales
}
\description{
Create a Labeller to Format Percentages with Icelandic Standards
}
