% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{label_isk}
\alias{label_isk}
\title{Create a Labeller to Format Numbers as Icelandic Kr<U+00F3>na}
\usage{
label_isk(scale = 1, suffix = NULL, accuracy = 1, ...)
}
\arguments{
\item{scale}{A scaling factor: x will be multiplied by scale before
formatting.}

\item{suffix}{If NULL the suffix will try to be decided based on
the scale.}

\item{accuracy}{A number to round to.}

\item{...}{Other inputs passed on to scales::label_number()}
}
\value{
A labelling function for use with ggplot2 scales
}
\description{
Create a Labeller to Format Numbers as Icelandic Kr<U+00F3>na
}
