% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{make_table_functions}
\alias{make_table_functions}
\title{Title}
\usage{
make_table_functions(table)
}
\arguments{
\item{table}{}
}
\value{
Creates a single helper function named \verb{mtl_<table>}
in the global environment.
}
\description{
Title
}
