% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{metill_db}
\alias{metill_db}
\title{R7 Class Representing a Connection to our database}
\usage{
metill_db(
  tables = class_missing,
  con_info = class_missing,
  connection = class_missing
)
}
\description{
R7 Class Representing a Connection to our database
}
