% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{metill_ggplotly}
\alias{metill_ggplotly}
\title{Convert ggplot2 to plotly with Metill styling}
\usage{
metill_ggplotly(plot, ...)
}
\arguments{
\item{plot}{A ggplot2 object to be converted to a plotly object}

\item{...}{Other inputs to ggplotly, layout or config from the plotly package}
}
\value{
A plotly object with transparent backgrounds, Lato font,
and hidden mode bar
}
\description{
Convert ggplot2 to plotly with Metill styling
}
