% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{scale_x_tufte}
\alias{scale_x_tufte}
\title{Make x, and y axes not touch (x axis part)}
\usage{
scale_x_tufte(...)
}
\value{
Apply to ggplot2 object
}
\description{
Make x, and y axes not touch (x axis part)
}
