% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{scale_y_tufte}
\alias{scale_y_tufte}
\title{Make x, and y axes not touch (y axis part)}
\usage{
scale_y_tufte(...)
}
\value{
Apply to ggplot2 object
}
\description{
Make x, and y axes not touch (y axis part)
}
