% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_visbending}
\alias{theme_visbending}
\title{Theme for the paper Vísbending}
\usage{
theme_visbending()
}
\value{
A theme object to be used with ggplot plots
}
\description{
Theme for the paper Vísbending
}
