% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stjornmalaflokkar.R
\name{translate_parties}
\alias{translate_parties}
\title{Title}
\usage{
translate_parties(name)
}
\arguments{
\item{name}{}
}
\value{
Character vector of translated party names
}
\description{
Title
}
