% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{tufte_breaks}
\alias{tufte_breaks}
\title{Helper function to breate axis breaks that include range values}
\usage{
tufte_breaks(x, n = 5, ...)
}
\arguments{
\item{x}{A vector containing the relevant data (x or y axis variable)}

\item{...}{}
}
\value{
A vector of breaks to be used by a ggplot2 scale
}
\description{
Helper function to breate axis breaks that include range values
}
