% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{hlutf}
\alias{hlutf}
\title{Format Numbers as Percentages with Icelandic Standards}
\usage{
hlutf(x, accuracy = 0.1, ...)
}
\arguments{
\item{x}{Numberic vector to be formatted as percentage}

\item{accuracy}{A number to round to. F.ex. 0.01 shows 2 decimal places.}
}
\value{
A character vector version of the input number, formatted as percentage with Icelandic localization
}
\description{
Format Numbers as Percentages with Icelandic Standards
}
