% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{isk}
\alias{isk}
\title{Format Numbers as Icelandic Krona}
\usage{
isk(x, scale = 1, suffix = NULL, ...)
}
\arguments{
\item{x}{The number to be formatted as ISK}

\item{scale}{A scaling factor: x will be multiplied by scale before formatting.}

\item{suffix}{If NULL the suffix will try to be decided based on the scale.}
}
\value{
A characted vector version of the input number formatted as ISK
}
\description{
Format Numbers as Icelandic Krona
}
