% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{label_hlutf}
\alias{label_hlutf}
\title{Create a Labeller to Format Numbers as Percentages with Icelandic Standards}
\usage{
label_hlutf(x, accuracy = 0.1, ...)
}
\arguments{
\item{x}{Numberic vector to be formatted as percentage}

\item{accuracy}{A number to round to. F.ex. 0.01 shows 2 decimal places.}

\item{...}{Other inputs passed on to scales::label_percent()}
}
\value{
A character vector version of the input number, formatted as percentage with Icelandic localization
}
\description{
Create a Labeller to Format Numbers as Percentages with Icelandic Standards
}
