% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{label_tala}
\alias{label_tala}
\title{Create a Labeller to Format Numbers with Icelandic Standards}
\usage{
label_tala(x, accuracy = 0.1, ...)
}
\arguments{
\item{x}{Numeric vector to be formatted as percentage}

\item{accuracy}{A number to round to. F.ex. 0.01 shows 2 decimal places.
#' @param ... Other inputs passed on to scales::label_number()}
}
\value{
A character vector version of the input number, formatted as percentage with Icelandic localization
}
\description{
Create a Labeller to Format Numbers with Icelandic Standards
}
