% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{metill_ggplotly}
\alias{metill_ggplotly}
\title{A customized ggplotly function to fit with Metill's visual theme}
\usage{
metill_ggplotly(plot, ...)
}
\arguments{
\item{plot}{A ggplot2 object to be converted to a plotly object}

\item{...}{Other inputs to ggplotly, layout or config from the plotly package}
}
\description{
A customized ggplotly function to fit with Metill's visual theme
}
