% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{tala}
\alias{tala}
\title{Format Numbers with Icelandic Standards}
\usage{
tala(x, accuracy = 0.1, ...)
}
\arguments{
\item{x}{Numberic vector to be formatted as percentage}

\item{accuracy}{A number to round to. F.ex. 0.01 shows 2 decimal places.}

\item{...}{Other inputs passed on to scales::number()}
}
\value{
A character vector version of the input number, formatted as percentage with Icelandic localization
}
\description{
Format Numbers with Icelandic Standards
}
